requirements.txt
python-telegram-bot==20.6
selenium==4.12.0


